sample/RAG/requirements.txt
FlagEmbedding
transformers
numpy
torch
faiss-cpu
OpenAI
pdfplumber
docx<=0.8.11 # for windows, where max_version is 0.2.x
streamlit
